requirements.txt
fastapi
pymongo
pydantic
pydantic[email]
python-decouple
uvicorn